+++ test/projects/svelte_ts_tailwind_eslint/eslint.config.js
@@ -1,0 +1,19 @@
+import pluginSvelte from "eslint-plugin-svelte";
+import tseslint from "typescript-eslint";
+
+export default [
+  ...tseslint.configs.recommended,
+  ...pluginSvelte.configs.recommended,
+  {
+    files: ["**/*.svelte"],
+    languageOptions: {
+      parserOptions: {
+        parser: tseslint.parser,
+        extraFileExtensions: [".svelte"],
+      },
+    },
+    rules: {
+      "@typescript-eslint/no-unused-vars": "error",
+    },
+  },
+];